backend/requirements.txt
fastapi
uvicorn
openai
python-multipart
groq
python-dotenv
sentence-transformers
numpy